Fix duplicate mappings, deprecated ansible variables and docker image inspect
