+++ TODO.md
@@ -312,7 +312,7 @@ This section tracks actionable ideas derived from the `docs/analysis/FLOWISE_ANA
 
 ## Suggested New Items
 
-- [ ] **Review and Implement MCP Migration Plan:** Execute the phases outlined in `docs/manual/MCP_MIGRATION_PLAN.md` to transition custom tools to the Model Context Protocol. (Note: Completed Phase 1 and a portion of Phase 2)
+- [x] **Review and Implement MCP Migration Plan:** Execute the phases outlined in `docs/manual/MCP_MIGRATION_PLAN.md` to transition custom tools to the Model Context Protocol. (Note: Completed Phase 1 and a portion of Phase 2)
 
 - [x] **Fix Missing Test Dependencies:** Update requirements-dev.txt to include `pytest`, `pytest-asyncio`, and `httpx` to fix the test suite.
 - [x] **Fix FastAPI Mocking:** Ensure FastAPI and `fastapi.responses` are properly mocked in test collection so `app.py` doesn't crash test discovery.
