+++ ansible.cfg
@@ -2,6 +2,9 @@
 force_valid_group_names = silently
 inject_facts_as_vars = False
 deprecation_warnings = False
+force_valid_group_names = silently
+inject_facts_as_vars = False
+deprecation_warnings = False
 collections_path = /root/.ansible/collections:/usr/share/ansible/collections
 fact_caching = jsonfile
 fact_caching_connection = /tmp/ansible_facts_cache
