+++ ansible/roles/consul/tasks/acl.yaml
@@ -79,9 +79,12 @@
 
         - name: Wait for Consul API to be ready after recovery
           uri:
-            url: "http://{{ advertise_ip }}:{{ consul_http_port }}/v1/status/leader"
+            url: "https://{{ advertise_ip }}:{{ consul_https_port }}/v1/status/leader"
+            client_cert: "{{ consul_config_dir | default('/etc/consul.d') }}/cert.pem"
+            client_key: "{{ consul_config_dir | default('/etc/consul.d') }}/key.pem"
+            validate_certs: false
           register: consul_status_recovery
-          until: consul_status_recovery.status == 200
+          until: consul_status_recovery.status is defined and consul_status_recovery.status == 200
           retries: 30
           delay: 5
