+++ ansible/roles/pipecatapp/tasks/main.yaml
@@ -402,33 +402,36 @@
   ansible.builtin.include_tasks:
     file: "{{ inventory_dir }}/ansible/tasks/build_pipecatapp_image.yaml"
 
-- name: Inspect pipecatapp:{{ pipecat_image_tag | default('local') }} image
-  ansible.builtin.command: docker image inspect pipecatapp:{{ pipecat_image_tag | default('local') }}
-  register: pipecatapp_local_image_info
-  changed_when: false
-  failed_when: false
-  ignore_errors: true
-  become: yes
+- name: Handle docker specific tasks for pipecatapp
+  block:
+    - name: Inspect pipecatapp:{{ pipecat_image_tag | default('local') }} image
+      ansible.builtin.command: docker image inspect pipecatapp:{{ pipecat_image_tag | default('local') }}
+      register: pipecatapp_local_image_info
+      changed_when: false
+      failed_when: false
+      ignore_errors: true
+      become: yes
 
-- name: Check if local registry is reachable (Pipecat App)
-  ansible.builtin.wait_for:
-    host: "{{ docker_registry_host }}"
-    port: "{{ docker_registry_port }}"
-    state: started
-    timeout: 1
-  register: pipecatapp_registry_check
-  failed_when: false
-  ignore_errors: true
+    - name: Check if local registry is reachable (Pipecat App)
+      ansible.builtin.wait_for:
+        host: "{{ docker_registry_host }}"
+        port: "{{ docker_registry_port }}"
+        state: started
+        timeout: 1
+      register: pipecatapp_registry_check
+      failed_when: false
+      ignore_errors: true
 
-- name: Set pipecatapp_image_ref (Registry)
-  ansible.builtin.set_fact:
-    pipecatapp_image_ref: "{{ docker_registry_host }}:{{ docker_registry_port }}/pipecatapp:{{ pipecat_image_tag | default('local') }}"
-  when: pipecatapp_registry_check.state | default('stopped') == 'started'
+    - name: Set pipecatapp_image_ref (Registry)
+      ansible.builtin.set_fact:
+        pipecatapp_image_ref: "{{ docker_registry_host }}:{{ docker_registry_port }}/pipecatapp:{{ pipecat_image_tag | default('local') }}"
+      when: pipecatapp_registry_check.state | default('stopped') == 'started'
 
-- name: Set pipecatapp_image_ref (Local Fallback)
-  ansible.builtin.set_fact:
-    pipecatapp_image_ref: "pipecatapp:{{ pipecat_image_tag | default('local') }}"
-  when: pipecatapp_registry_check.state | default('stopped') != 'started'
+    - name: Set pipecatapp_image_ref (Local Fallback)
+      ansible.builtin.set_fact:
+        pipecatapp_image_ref: "pipecatapp:{{ pipecat_image_tag | default('local') }}"
+      when: pipecatapp_registry_check.state | default('stopped') != 'started'
+  when: pipecat_deployment_style == 'docker'
 
 - name: Ensure Nomad jobs directory exists
   ansible.builtin.file:
