+++ tests/unit/test_mcp_tool.py
@@ -4,9 +4,9 @@ import os
 from unittest.mock import MagicMock
 
 # Add tools directory to path
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'ansible', 'roles', 'pipecatapp', 'files', 'tools')))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..', 'pipecatapp')))
 
-from mcp_tool import MCP_Tool
+from tools.mcp_tool import MCP_Tool
 
 @pytest.fixture
 def mcp_tool():
